Corrected a coordinate problem with the shading.
Shading in PDF is using the target coordinates. So, if a diagram
is transformed (translated, rotated, scaled) the shading must also be
transformed. But, we cannot just use the standard matrix of transformation
for the geometry since the shading coordinates are saved in a resource
section separated from the drawing commands.

In withStyle we only have access to the transform after freezing. But, all
transforms are needed. So a withStyle' is introduced and use only for the
renderDia implementation.

With means that the standard withStyle is not fully supporting shading and
in its current form can't.